+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ __Changes__
 - updated BLAST to 2.13.0
 - fixed bug with long feature IDs
 - added SpeciesTree as possible target
+- add option to write off-genetic-code protein translations
 
 ### Version 1.6.0
 __Changes__